AI/requirements.txt
python-dotenv>=0.19.0
openai>=0.27.0
requests>=2.26.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pytest>=6.2.5
tqdm>=4.62.0
